twizzler_rt_abi::alloc

Function twz_rt_dealloc

pub unsafe fn twz_rt_dealloc(ptr: *mut u8, layout: Layout, flags: AllocFlags)

Deallocate runtime memory pointed to by ptr, with a given layout and flags.

Safety

Caller must ensure that any no other references to this memory are alive.

Flags

ZERO_MEMORY: Zero the old memory before freeing.