twizzler_rt_abi::core

Module rt0

rt0 defines a collection of functions that the basic Rust ABI expects to be defined by some part of the C runtime:

__tls_get_addr for handling non-local TLS regions.

Functions

__rust_entry_from_c⚠

rust_entry⚠

Entry point for Rust code wishing to start from rt0.