twizzler_rt_abi

Module object

Interface for objects and object handles.

Structs

CreateTieFlags

Flags controlling how a particular object tie operates.

CreateTieSpec

A specification of ties to create. (see the book for more information on ties).

FotEntry

FotFlags

MapFlags

Flags for mapping objects.

MetaExt

A meta extension entry.

MetaExtTag

A tag for a meta extension entry.

MetaFlags

Flags for objects.

MetaInfo

The core metadata that all objects share.

Nonce

A nonce for avoiding object ID collision.

ObjID

An object ID.

ObjectCreate

Full object creation specification, minus ties.

ObjectCreateFlags

Flags to pass to the object create system call.

ObjectHandle

An object handle, granting access to object memory. An object handle can be in two modes:

ObjectSource

Specifications for an object-copy from a source object. The specified ranges are source:[src_start, src_start + len) copied to <some unspecified destination object>:[dest_start, dest_start + len). Each range must start within an object, and end within the object.

Protections

Mapping protections for mapping objects into the address space.

Enums

BackingType

The backing memory type for this object. Currently doesn’t do anything.

LifetimeType

The base lifetime type of the object. Note that this does not ensure that the object is stored in a specific type of memory, the kernel is allowed to migrate objects with the Normal BackingType as it sees fit. For more information on object lifetime, see the book.

ObjectCmd

Constants

MAX_SIZE

The maximum size of an object, including null page and meta page(s).

MEXT_EMPTY

MEXT_SIZED

NULLPAGE_SIZE

The size of the null page.

Functions

twz_rt_get_object_handle

twz_rt_insert_fot

twz_rt_map_object

Map an object given by ID id with the given flags.

twz_rt_map_two_objectsDeprecated

twz_rt_release_handle

Release a handle. Should be only called by the ObjectHandle drop call.

twz_rt_resolve_fot

twz_rt_resolve_fot_local

twz_rt_update_handle

Update a handle.