twizzler_rt_abi
Module bindings

Structs

basic_aux

Basic OS information provided to rust libstd

basic_return

Return info from rust libstd

comp_init_info

Init information for compartments

dl_phdr_info

Information about a loaded dynamic object

duration

Duration, containing seconds and nanoseconds.

fd_info

Information about a file descriptor.

io_result

Result of IO operations

io_vec

Io vec, a buffer and a len.

loaded_image

Information about a loaded program image or library

map_result

Result map_object call

max_align_t

minimal_init_info

Init information for minimal runtime

object_handle

Object handle

open_info

Information for opening a file.

open_result

Result of open call.

option_duration

Optional duration.

option_exit_code

Optional exit code

runtime_info

Runtime initialization info.

spawn_args

Arguments to spawn

spawn_result

Spawn result.

system_info

Information about the system

tls_index

TLS index, module ID and offset.

Constants

FD_CMD_DUP

Duplicate this descriptor. The arg argument is ignored. The ret argument points to a descriptor.

FD_CMD_SUCCESS

Success value for twz_rt_fd_cmd.

FD_IS_TERMINAL

This file descriptor is a terminal.

FD_POS

FUTEX_WAKE_ALL

Wake all threads instead of a maximum number

GET_RANDOM_NON_BLOCKING

Do not block when collecting random data

IO_NONBLOCKING

Non-blocking behavior specified. If the operation would block, return io_result with error set to WouldBlock instead.

LEN_MUL

Multiplier to valid_len.

MAP_FLAG_R

Map with READ permission.

MAP_FLAG_W

Map with WRITE permission.

MAP_FLAG_X

Map with EXEC permission.

RUNTIME_INIT_COMP

Info for compartments

RUNTIME_INIT_MIN

Minimal runtime info

RUNTIME_INIT_MONITOR

Info for monitor

TWZ_RT_EXEID

The loaded image ID for the root loaded image (usually the executable)

WHENCE_CURRENT

Seek offset from current fd position

WHENCE_END

Seek offset from end of file

WHENCE_START

Seek offset from start of file

ZERO_MEMORY

Zero memory during operation

__bool_true_false_are_defined

false_

io_error_IoError_InvalidDesc

Invalid descriptor

io_error_IoError_Other

Other error

io_error_IoError_SeekError

Error during seek

io_error_IoError_Success

Success

io_error_IoError_WouldBlock

Operation would block, but nonblocking behavior was specified.

join_result_Join_Other

join_result_Join_Success

join_result_Join_ThreadNotFound

join_result_Join_Timeout

map_error_MapError_InvalidArgument

map_error_MapError_NoSuchObject

map_error_MapError_Other

map_error_MapError_OutOfResources

map_error_MapError_PermissionDenied

map_error_MapError_Success

monotonicity_NonMonotonic

monotonicity_StrongMonotonic

monotonicity_WeakMonotonic

open_error_OpenError_InvalidArgument

Argument was invalid.

open_error_OpenError_LookupFail

Failed to lookup provided file name.

open_error_OpenError_Other

Other error.

open_error_OpenError_PermissionDenied

Permission denied.

open_error_OpenError_Success

Open success.

spawn_error_Spawn_InvalidArgument

spawn_error_Spawn_KernelError

spawn_error_Spawn_ObjectNotFound

spawn_error_Spawn_Other

spawn_error_Spawn_PermissionDenied

spawn_error_Spawn_Success

true_

version

Functions

__twz_rt_map_two_objects⚠

twz_rt_abort⚠

Abort immediately

twz_rt_dealloc⚠

Deallocate memory. If ZERO_MEMORY is set, will clear the memory before freeing.

twz_rt_exit⚠

Exit with provided code

twz_rt_fd_close⚠

Close a file descriptor. If the file descriptor is invalid or already closed, this function does nothing.

twz_rt_fd_cmd⚠

Perform a command on the descriptor. The arguments arg and ret are interpreted according to the command specified.

twz_rt_fd_get_info⚠

Get information about a descriptor. If this returns true, the fd was valid and the data pointed to by info is filled with fd_info data.

twz_rt_fd_open⚠

Open a file.

twz_rt_fd_pread⚠

Read from a file. May read less than specified len.

twz_rt_fd_preadv⚠

Do vectored IO read.

twz_rt_fd_pwrite⚠

Write to a file. May write less than specified len.

twz_rt_fd_pwritev⚠

Do vectored IO write.

twz_rt_fd_seek⚠

Seek to a specified point in the file.

twz_rt_futex_wait⚠

If *ptr == expected, wait until signal, optionally timing out.

twz_rt_futex_wake⚠

Wake up up to max threads waiting on ptr. If max is set to FUTEX_WAKE_ALL, wake all threads.

twz_rt_get_loaded_image⚠

Get a loaded image from its ID. All IDs for loaded image are sequential, starting from TWZ_RT_EXEID. On success, fill out data pointed to by the li argument and return true.

twz_rt_get_monotonic_time⚠

Get time from the monotonic clock

twz_rt_get_random⚠

Collect up to len bytes of randomness, filling buf. Returns the number of bytes of random data actually collected.

twz_rt_get_sysinfo⚠

Get system information

twz_rt_get_system_time⚠

Get time from the system clock

twz_rt_iter_phdr⚠

twz_rt_join_thread⚠

Wait for a thread to exit, optionally timing out.

twz_rt_malloc⚠

Allocate memory, zeroing it first if the flag is set.

twz_rt_map_object⚠

Map an object with a given ID and flags.

twz_rt_post_main_hook⚠

twz_rt_pre_main_hook⚠

twz_rt_realloc⚠

Reallocate memory. If ZERO_MEMORY is set, will zero new memory before returning and zero to-be-freed memory before freeing.

twz_rt_release_handle⚠

Release an object handle. After calling this, the handle may not be used.

twz_rt_runtime_entry⚠

twz_rt_set_name⚠

Set the name of the calling thread.

twz_rt_sleep⚠

Sleep the calling thread for specified duration.

twz_rt_spawn_thread⚠

Sawn a thread. On success, that thread starts executing concurrently with this function’s return.

twz_rt_tls_get_addr⚠

Resolve the TLS index and get back the TLS data pointer.

twz_rt_yield_now⚠

Yield the thread now.

Type Aliases

__uint128_t

Object ID

alloc_flags

Allocation flags

descriptor

An open descriptor for a runtime file handle.

exit_code

Exit code type

fd_cmd

Commands for descriptors.

fd_cmd_err

Errors for twz_rt_fd_cmd.

fd_flags

Flags a descriptor can have.

futex_word

Futex type, based on linux futex.

get_random_flags

Flags to get_random

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

io_error

Possible IO errors

io_flags

Flags for IO operations

join_result

Possible results of join.

loaded_image_id

An ID for a loaded program image (or library)

map_error

Possible mapping errors

map_flags

Mapping flags

monotonicity

Supported monotonicity levels

open_error

Possible open error conditions.

optional_offset

Optional offset. If value is FD_POS, use the file descriptor position.

rt_objid

Object ID

spawn_error

Possible spawn errors

thread_id

Runtime-internal ID of a thread

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t

whence

Type of whence values for seek.

Unions

init_info_ptrs

Possible init info types