twizzler_rt_abi

Module core

Low-level runtime functionality.

Re-exports

pub use crate::bindings::basic_aux as BasicAux;

pub use crate::bindings::basic_return as BasicReturn;

pub use crate::bindings::comp_init_info as CompartmentInitInfo;

pub use crate::bindings::ctor_set as CtorSet;

pub use crate::bindings::init_info_ptrs as InitInfoPtrs;

pub use crate::bindings::minimal_init_info as MinimalInitInfo;

pub use crate::bindings::runtime_info as RuntimeInfo;

pub use crate::bindings::RUNTIME_INIT_COMP;

pub use crate::bindings::RUNTIME_INIT_MIN;

pub use crate::bindings::RUNTIME_INIT_MONITOR;

Modules

rt0

rt0 defines a collection of functions that the basic Rust ABI expects to be defined by some part of the C runtime:

Functions

twz_rt_abort

Abort execution due to unrecoverable language error.

twz_rt_cross_compartment_entry

Called by security context code on compartment entry

twz_rt_exit

Exit with the provided error code. If the main thread for a program exits, the remaining threads will exit as well.

twz_rt_post_main_hook

Call this after return from main, before running destructors.

twz_rt_pre_main_hook

Call this before calling main, after initializing the runtime. If this function returns None, then call main. Otherwise, act as if main returned the provided ExitCode.

twz_rt_runtime_entry

The entry point for the runtime. Not for public use.

Type Aliases

ExitCode

Type for exit code.