twizzler_rt_abi

Module fd

Runtime interface for file descriptors.

Re-exports

pub use crate::bindings::descriptor as RawFd;

pub use crate::bindings::name_entry as NameEntry;

Structs

FdFlags

Flags for file descriptors.

FdInfo

Information about an open file descriptor.

SocketAddress

Enums

FdKind

Possible Fd Kinds

OpenAnonKind

Functions

twz_rt_fd_close

Close a file descriptor. If the fd is already closed, or invalid, this function has no effect.

twz_rt_fd_copen

Open a file descriptor by name, as a C-string.

twz_rt_fd_dup

Duplicate a file descriptor.

twz_rt_fd_enumerate_names

Enumerate sub-names for an fd (e.g. directory entries). Returns Some(n) on success, None if no names can be enumerated. Return of Some(n) indicates number of items read into the buffer, 0 if end of name list. Offset argument specifies number of entries to skip.

twz_rt_fd_get_info

Get information about an open file descriptor. If the fd is invalid or closed, returns None.

twz_rt_fd_mkns

Make a new namespace

twz_rt_fd_open

Open a file descriptor by name, as a Rust-string.

twz_rt_fd_open_pipe

Open an anonymous file descriptor.

twz_rt_fd_open_socket_bind

Open an anonymous file descriptor.

twz_rt_fd_open_socket_connect

Open an anonymous file descriptor.

twz_rt_fd_readlink

twz_rt_fd_remove

Remove a name

twz_rt_fd_symlink

Make a new symlink

twz_rt_fd_sync

Sync a file descriptor.

twz_rt_fd_truncate

Truncate a file descriptor.